requirements.txt
Flask
Flask-login
Flask-SQLAlchemy
python-dotenv
Werkzeug
gunicorn

